Move test classes to correct namespace
